Move non-user code to lib
