+++ template/src/components/layers/StoresLayer.js
@@ -1,9 +1,8 @@
 import { useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import { CartoSQLLayer } from '@deck.gl/carto';
-import { getFilteredQuery } from 'lib/sdk';
+import { getFilteredQuery, currencyFormatter } from 'lib/sdk';
 import { selectSourceById } from 'config/cartoSlice';
-import { currencyFormatter } from 'utils/numberFormatters';
 
 export const LayerStyle = {
   id: 'storesLayer',
