+++ template/src/components/legends/Legend.js
