+++ template/src/lib/sdk/api/index.js
@@ -1,0 +1,5 @@
+import { executeSQL } from './SQL';
+import { getDatasets } from './Datasets';
+import { dealWithApiError, generateApiUrl } from './common';
+
+export { executeSQL, getDatasets, dealWithApiError, generateApiUrl };
