+++ template/src/lib/sdk/google/index.js
@@ -1,0 +1,2 @@
+import { GoogleMap } from './GoogleMap';
+export { GoogleMap };
