+++ template/src/lib/sdk/index.js
@@ -1,24 +1,5 @@
-import { executeSQL } from './SQL';
-import { getDatasets } from './Datasets';
-
-import { AggregationTypes } from './models/AggregationTypes';
-import { getCategories } from './models/CategoryModel';
-import { getValue } from './models/FormulaModel';
-import {
-  FilterTypes,
-  getFilterCondition,
-  getConditionFromViewPort,
-  getFilteredQuery,
-} from './models/FilterConditionBuilder';
-
-export {
-  executeSQL,
-  getDatasets,
-  AggregationTypes,
-  FilterTypes,
-  getCategories,
-  getValue,
-  getFilterCondition,
-  getConditionFromViewPort,
-  getFilteredQuery,
-};
+export * from './api';
+export * from './widgets';
+export * from './utils';
+export * from './google';
+export * from './models';
