+++ template/src/lib/sdk/models/index.js
@@ -1,0 +1,19 @@
+import { AggregationTypes } from './AggregationTypes';
+import { getCategories } from './CategoryModel';
+import { getValue } from './FormulaModel';
+import {
+  FilterTypes,
+  getFilterCondition,
+  getConditionFromViewPort,
+  getFilteredQuery,
+} from './FilterConditionBuilder';
+
+export {
+  AggregationTypes,
+  FilterTypes,
+  getCategories,
+  getValue,
+  getFilterCondition,
+  getConditionFromViewPort,
+  getFilteredQuery,
+};
