+++ template/src/lib/sdk/widgets/CategoryWidget.js
